+++ plugins/main/common/constants.ts
@@ -1134,3 +1134,9 @@ export const VULNERABILITIES_DASHBOARD_ID =
   'vulnerabilities-overview-dashboard-tab';
 export const VULNERABILITIES_AGENT_DASHBOARD_ID =
   'vulnerabilities-overview-dashboard-tab';
+
+// Main module tabs
+export const TAB_VIEW_ID_EVENTS = 'findings';
+export const TAB_VIEW_NAME_EVENTS = 'Findings';
+export const TAB_VIEW_ID_DASHBOARD = 'dashboard';
+export const TAB_VIEW_NAME_DASHBOARD = 'Dashboard';